+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-14T19:19:01.245Z",
+    "2026-05-15T03:37:24.790Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {