+++ packages/orq-rc/RELEASES.md
@@ -9209,3 +9209,13 @@ Based on:
 - [typescript v4.10.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.6 - packages/orq-rc
+
+## 2026-05-15 03:40:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.7 - packages/orq-rc